Add two new spatial fields
There are now two new spatial fields to identify the geographical
coverage more specifically. The spatial-code will be used as a
standardized identifier that consists of a NUTS code and the
Amtlicher Gemeindeschlüssel which is separated by a slash.

The field spatial-uri is used to introduce linked data to the
metadata schema which makes it possible link resources which
embody the geographical coverage.

Signed-off-by: Konrad Reiche <konrad.reiche@fokus.fraunhofer.de>