Fix coordinates specification
The coordinates specification for the extras field spatial described an
array of arrays of numbers. According to the GeoJSON specification,
however, it has to be an array of arrays of arrays of numbers. This is
due to the fact, that the polygon needs to be a list of linear ring,
which is line string that closes itself in the end again.

Signed-off-by: Konrad Reiche <konrad.reiche@fokus.fraunhofer.de>