Add new field to prevent duplicates in the future
A new field is added to prevent possible duplicates which can result
from harvesting other endpoints but CKAN. For instance, a CSV harvester
has to perform a metadata mapping. If different portals do that it
becomes unclear who 'owns' the dataset and which portal should be
prioritized when harvesting these duplicates. See #17.

Signed-off-by: Konrad Reiche <konrad.reiche@fokus.fraunhofer.de>